+++ lisp/eab-desktop.el
@@ -9,7 +9,7 @@
 
 (require 'desktop)
 
-(if (or (eab/ondaemon "serverP") (eab/ondaemon "serverC") (eab/ondaemon "serverM"))
+(if (or (eab/ondaemon (eab/server-P)) (eab/ondaemon (eab/server-C)) (eab/ondaemon "serverM"))
     (add-hook 'desktop-save-hook (lambda ()
 				   ;; (eab/wg-update-all-workgroups)
 				   (eab/wg-save eab/workgroups-save)