+++ lisp/eab-tex.el
@@ -25,7 +25,7 @@
 (setq TeX-parse-self t)
 (setq TeX-PDF-mode t)
 (setq-default TeX-master 't)
-(if (not (eab/ondaemon "serverP")) (setq-default TeX-master 't))
+(if (not (eab/ondaemon (eab/server-P))) (setq-default TeX-master 't))
 (setq TeX-open-quote "<<")
 (setq TeX-close-quote ">>")
 ;; (load "preview-latex.el" nil t t)